[refactor] Modal 컴포넌트들 리팩토링
#157
