+++ src/components/modals/AcceptModal/hooks/useAnimation.ts
