+++ src/components/modals/AcceptModal/index.tsx
