+++ src/components/modals/AcceptModal/reducer.ts
