+++ src/components/modals/AcceptModal/styled.ts
