+++ src/components/modals/CopyModal/styled.ts
@@ -1,0 +1,16 @@
+import styled from 'styled-components/native';
+
+import Text from '@src/components/Text';
+
+export const Wrapper = styled.View`
+  padding: 20px;
+`;
+
+export const CopyButtonBody = styled.View`
+  flex-direction: row;
+  align-items: center;
+`;
+
+export const CopyButtonTitle = styled(Text)`
+  margin-left: 4px;
+`;
