+++ src/components/modals/DefaultModal/styled.ts
@@ -10,13 +10,14 @@ export const ModalContainer = styled.View`
 
 export const Overlay = styled.TouchableOpacity`
   ${positionAbsoluteFill};
-  opacity: 0.4;
+  opacity: 0.6;
   background-color: ${({ theme }) => theme.foreground};
 `;
 
 export const Container = styled.View`
-  width: 95%;
-  padding: 20px;
+  width: 90%;
+  border-radius: 8px;
+  overflow: hidden;
   background-color: ${({ theme }) => theme.background};
 `;
 
@@ -24,16 +25,5 @@ export const CloseButton = styled(Button)`
   position: absolute;
   top: 5px;
   right: 5px;
-`;
-
-export const FlatListWrapper = styled.View`
-  margin-top: 20px;
-`;
-
-export const StyledFlatList = styled.FlatList`
-  max-height: 500px;
-`;
-
-export const LoadButton = styled(Button)`
-  width: 95%;
+  z-index: 10;
 `;
