+++ src/components/modals/PreviousPlansModal/styled.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components/native';
+
+export const Container = styled.View`
+  padding: 20px;
+`;
+
+export const FlatListWrapper = styled.View`
+  margin-top: 20px;
+`;
+
+export const StyledFlatList = styled.FlatList`
+  max-height: 500px;
+`;
