+++ src/components/modals/TrainingModal/styled.ts
@@ -1,30 +1,16 @@
-import { transparentize } from 'polished';
 import styled from 'styled-components/native';
 
 import Button from '@src/components/Button';
 import Image from '@src/components/Image';
-import { flexFillCenter } from '@src/styles/flex';
-
-export const Overlay = styled(Button)`
-  background-color: ${({ theme }) => transparentize(0.6, theme.foreground)};
-  ${flexFillCenter};
-`;
 
 export const Container = styled.View`
   padding: 20px;
-  background-color: ${({ theme }) => theme.background};
-`;
-
-export const CloseButton = styled(Button)`
-  position: absolute;
-  right: 5px;
-  top: 5px;
 `;
 
 export const Thumbnail = styled(Image)`
   margin-top: 10px;
-  width: 240px;
-  height: 240px;
+  width: 100%;
+  height: 280px;
 `;
 
 export const CategoryContainer = styled.View`
